[PAGOPA-1362] refactoring function
