+++ src/main/java/it/gov/pagopa/gpd/upload/UploadFunction.java
@@ -1,6 +1,5 @@
 package it.gov.pagopa.gpd.upload;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.microsoft.azure.functions.ExecutionContext;
@@ -10,13 +9,11 @@ import com.microsoft.azure.functions.annotation.*;
 import it.gov.pagopa.gpd.upload.entity.FailedIUPD;
 import it.gov.pagopa.gpd.upload.model.ResponseGPD;
 import it.gov.pagopa.gpd.upload.entity.Status;
-import it.gov.pagopa.gpd.upload.entity.Upload;
-import it.gov.pagopa.gpd.upload.exception.AppException;
-import it.gov.pagopa.gpd.upload.model.RetryStep;
 import it.gov.pagopa.gpd.upload.model.pd.PaymentPositionModel;
 import it.gov.pagopa.gpd.upload.model.pd.PaymentPositionsModel;
 import it.gov.pagopa.gpd.upload.client.GpdClient;
 import it.gov.pagopa.gpd.upload.repository.StatusRepository;
+import it.gov.pagopa.gpd.upload.service.StatusService;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
@@ -35,8 +32,6 @@ import java.util.stream.Collectors;
 public class UploadFunction {
 
     private final String BLOCK_SIZE = System.getenv("BLOCK_SIZE");
-
-    private final int MESSAGE_MAX_CHAR_NUMBER = 150;
     private final int BAD_REQUEST = 400;
 
     /**
@@ -72,17 +67,15 @@ public class UploadFunction {
         try {
             // deserialize payment positions from JSON to Object
             PaymentPositionsModel pps = objectMapper.readValue(converted, PaymentPositionsModel.class);
-            Status status = this.createStatus(logger, fiscalCode, key, pps);
+            Status status = StatusService.getInstance(logger).createStatus(fiscalCode, key, pps);
             logger.log(Level.INFO, () -> "Payment positions size: " + pps.getPaymentPositions().size());
             // function logic: validation and block upload to GPD-Core
             validate(logger, pps, status);
             createPaymentPositionBlocks(logger, context.getInvocationId(), fiscalCode, key, pps, status);
             // write status in output container
             outputBlob.setValue(objectMapper.writeValueAsString(status));
-        } catch (JsonProcessingException | AppException e) {
-            logger.log(Level.INFO, () -> "Processing blob exception: " + e.getMessage());
         } catch (Exception e) {
-            logger.log(Level.INFO, () -> "Processing blob exception: " + e);
+            logger.log(Level.INFO, () -> "Processing function exception: " + e.getMessage());
         }
     }
 
@@ -100,7 +93,7 @@ public class UploadFunction {
             block = new PaymentPositionsModel(pps.getPaymentPositions().subList(index, index+blockSize));
             ResponseGPD response = GpdClient.getInstance().createDebtPositions(fc, block, logger, invocationId);
             List<String> IUPDs = block.getPaymentPositions().stream().map(item -> item.getIupd()).collect(Collectors.toList());
-            this.updateStatus(IUPDs, status, response, blockSize);
+            StatusService.getInstance(logger).updateStatus(IUPDs, status, response, blockSize);
             StatusRepository.getInstance(logger).upsertStatus(key, status);
             index += blockSize;
         }
@@ -112,7 +105,7 @@ public class UploadFunction {
             block = new PaymentPositionsModel(pps.getPaymentPositions().subList(index, index+remainingPosition));
             ResponseGPD response = GpdClient.getInstance().createDebtPositions(fc, block, logger, invocationId);
             List<String> IUPDs = block.getPaymentPositions().stream().map(pp -> pp.getIupd()).collect(Collectors.toList());
-            this.updateStatus(IUPDs, status, response, remainingPosition);
+            StatusService.getInstance(logger).updateStatus(IUPDs, status, response, remainingPosition);
             StatusRepository.getInstance(logger).upsertStatus(key, status);
         }
         if(status.upload.getCurrent() == status.upload.getTotal()) {
@@ -158,44 +151,5 @@ public class UploadFunction {
 
         status.upload.setCurrent(status.upload.getCurrent() + failedIUPDs.size());
         status.upload.setFailedIUPDs(failedIUPDs);
-    }
-
-    private Status createStatus(Logger logger, String fiscalCode, String key, PaymentPositionsModel paymentPositionsModel) throws AppException {
-        Status statusIfNotExist = Status.builder()
-                                          .id(key)
-                                          .fiscalCode(fiscalCode)
-                                          .upload(Upload.builder()
-                                                          .current(0)
-                                                          .total(paymentPositionsModel.getPaymentPositions().size())
-                                                          .successIUPD(new ArrayList<>())
-                                                          .failedIUPDs(new ArrayList<>())
-                                                          .start(LocalDateTime.now()).build())
-                                          .build();
-        Status status = StatusRepository.getInstance(logger).createIfNotExist(key, fiscalCode, statusIfNotExist);
-        if(status.upload.getEnd() != null) {
-            logger.log(Level.INFO, () -> "Upload already processed. Upload finished at " + status.upload.getEnd());
-            return status;
-        }
-
-        return status;
-    }
-
-    public Status updateStatus(List<String> IUPDs, Status status, ResponseGPD response, int blockSize) {
-        RetryStep responseRetryStep = response.getRetryStep();
-        if(responseRetryStep.equals(RetryStep.DONE)) {
-            ArrayList<String> successIUPDs = status.upload.getSuccessIUPD();
-            successIUPDs.addAll(IUPDs);
-            status.upload.setSuccessIUPD(successIUPDs);
-        } else if(responseRetryStep.equals(RetryStep.ERROR) || responseRetryStep.equals(RetryStep.RETRY) || responseRetryStep.equals(RetryStep.NONE)  ) {
-            FailedIUPD failedIUPD = FailedIUPD.builder()
-                    .details(response.getDetail().substring(0, Math.min(response.getDetail().length(), MESSAGE_MAX_CHAR_NUMBER)))
-                    .errorCode(response.getStatus())
-                    .skippedIUPDs(IUPDs)
-                    .build();
-            status.upload.addFailures(failedIUPD);
-        }
-        status.upload.setCurrent(status.upload.getCurrent() + blockSize);
-
-        return status;
     }
 }
