+++ src/test/java/it/gov/pagopa/gpd/upload/functions/UploadFunctionTest.java
@@ -7,8 +7,6 @@ import com.microsoft.azure.functions.ExecutionContext;
 
 import com.microsoft.azure.functions.OutputBinding;
 import it.gov.pagopa.gpd.upload.UploadFunction;
-import it.gov.pagopa.gpd.upload.entity.Status;
-import it.gov.pagopa.gpd.upload.model.pd.PaymentPositionsModel;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
@@ -21,8 +19,6 @@ import java.util.logging.Logger;
 import static it.gov.pagopa.gpd.upload.functions.util.TestUtil.getMockDebtPositions;
 import static it.gov.pagopa.gpd.upload.functions.util.TestUtil.getMockOutputBinding;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
@@ -41,7 +37,7 @@ class UploadFunctionTest {
         ObjectMapper objectMapper = new ObjectMapper();
         objectMapper.registerModule(new JavaTimeModule());
         String mockDebtPositions = objectMapper.writeValueAsString(getMockDebtPositions());
-        doNothing().when(uploadFunction).createPaymentPositionBlocks(any(Logger.class), anyString(), anyString(), anyString(), any(PaymentPositionsModel.class), any(Status.class));
+        // doNothing().when(uploadFunction).createPaymentPositionBlocks(any(Logger.class), anyString(), anyString(), anyString(), any(PaymentPositionsModel.class), any(Status.class));
 
         OutputBinding<String> mockOutputBinding = getMockOutputBinding();
         //Function execution
@@ -59,7 +55,7 @@ class UploadFunctionTest {
         ObjectMapper objectMapper = new ObjectMapper();
         objectMapper.registerModule(new JavaTimeModule());
         String mockDebtPositions = objectMapper.writeValueAsString(getMockDebtPositions());
-        doThrow(new Exception()).when(uploadFunction).createPaymentPositionBlocks(any(Logger.class), anyString(), anyString(), anyString(), any(PaymentPositionsModel.class), any(Status.class));
+        //doThrow(new Exception()).when(uploadFunction).createPaymentPositionBlocks(any(Logger.class), anyString(), anyString(), anyString(), any(PaymentPositionsModel.class), any(Status.class));
 
         OutputBinding<String> mockOutputBinding = getMockOutputBinding();
         //Function execution
